requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.5
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
Pillow==10.2.0
python-telegram-bot==20.8 